requirements.txt
asgiref==3.5.0
Django==4.0.2
gunicorn==20.1.0
mysql==0.0.3
mysqlclient==2.1.0
numpy
pandas
Pillow==9.0.1
py==1.11.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==5.3.0
matplotlib
scikit-learn
lime
joblib
import dill